fix(spatialflow): compile — Float elvis fallback widened percentDownloaded to Number&Comparable
media3's Download.percentDownloaded is a platform Float; '?: 0.0' (Double
literal) makes the elvis type the LUB intersection Number&Comparable<*>,
which no maxOf overload accepts (spatial flow download chip progress).
Keep the fallback Float and convert to Double explicitly, matching
fetchPercent.toDouble().

+++ app/src/main/kotlin/moe/rukamori/archivetune/ui/player/spatialflow/SpatialFlowPlayer.kt
@@ -523,7 +523,11 @@ fun SpatialFlowPlayerContent(
                             // temp file Media3 reports 0% — combine in the live
                             // fetch progress so the label reflects the network
                             // download instead of a fake "Downloading 0%".
-                            val media3Percent = download?.percentDownloaded ?: 0.0
+                            // media3's percentDownloaded is a Java float; the
+                            // elvis must stay Float (a 0.0 Double fallback
+                            // widens the type to Number&Comparable, which no
+                            // maxOf overload accepts).
+                            val media3Percent = (download?.percentDownloaded ?: 0f).toDouble()
                             val fetchPercent = fetchProgressMap[downloadUtil
                                 .currentSourceDownloadTarget(mediaMetadata.id)
                                 .key]?.percent ?: 0
